+++ app/src/main/java/com/fibelatti/photowidget/model/PhotoWidgetTapAction.kt
@@ -10,5 +10,6 @@ enum class PhotoWidgetTapAction(
     NONE(title = R.string.photo_widget_configure_tap_action_none),
     VIEW_FULL_SCREEN(title = R.string.photo_widget_configure_tap_action_view_full_screen),
     VIEW_NEXT_PHOTO(title = R.string.photo_widget_configure_tap_action_next_photo),
+    APP_SHORTCUT(title = R.string.photo_widget_configure_tap_action_app_shortcut),
     ;
 }